Refactor OpSemMemManager into a virtual interface and implementation
