+++ lib/seahorn/BvOpSem2Context.hh
@@ -109,7 +109,7 @@ public:
   Bv2OpSemContext(const Bv2OpSemContext &) = delete;
   ~Bv2OpSemContext() override = default;
 
-  EZ3* getZ3() const {return m_z3.get();}
+  EZ3 *getZ3() const { return m_z3.get(); }
 
   /// \brief Writes value \p u into symbolic register \p v
   void write(Expr v, Expr u);
@@ -433,19 +433,13 @@ std::unique_ptr<OpSemAllocator> mkStaticOpSemAllocator(OpSemMemManager &mem);
 
 /// \brief Memory manager for OpSem machine
 class OpSemMemManager {
-
+protected:
   /// \brief Parent Operational Semantics
   Bv2OpSem &m_sem;
   /// \brief Parent Semantics Context
   Bv2OpSemContext &m_ctx;
   /// \brief Parent expression factory
   ExprFactory &m_efac;
-
-  /// \brief Allocates memory regions in virtual memory
-  std::unique_ptr<OpSemAllocator> m_allocator;
-
-  /// \brief Knows the memory representation and how to access it
-  std::unique_ptr<OpSemMemRepr> m_memRepr;
 
   /// \brief Ptr size in bytes
   uint32_t m_ptrSz;
@@ -456,24 +450,11 @@ class OpSemMemManager {
   /// Must be divisible by \t m_wordSz
   uint32_t m_alignment;
 
-  /// \brief Base name for non-deterministic pointer
-  Expr m_freshPtrName;
-
-  /// \brief Register that contains the value of the stack pointer on
-  /// function entry
-  Expr m_sp0;
-
-  /// \brief Source of unique identifiers
-  mutable unsigned m_id;
-
-  /// \brief A null pointer expression (cache)
-  Expr m_nullPtr;
-
 public:
   OpSemMemManager(Bv2OpSem &sem, Bv2OpSemContext &ctx, unsigned ptrSz,
-                  unsigned wordSz, bool useLambdas = false);
+                  unsigned wordSz);
 
-  ~OpSemMemManager() = default;
+  virtual ~OpSemMemManager() = default;
 
   /// Right now everything is an expression. In the future, we might have other
   /// types for PtrTy, such as a tuple of expressions
@@ -483,74 +464,74 @@ public:
   Bv2OpSemContext &ctx() const { return m_ctx; }
 
   unsigned ptrSzInBits() const { return m_ptrSz * 8; }
-  PtrTy ptrSort() const { return bv::bvsort(ptrSzInBits(), m_efac); }
+  unsigned wordSzInBytes()const {return m_wordSz;}
+  unsigned wordSzInBits()const {return m_wordSz * 8;}
 
-  unsigned wordSzInBytes() const { return m_wordSz; }
-  unsigned wordSzInBits() const { return m_wordSz * 8; }
+  virtual PtrTy ptrSort() const = 0;
 
   /// \brief Allocates memory on the stack and returns a pointer to it
   /// \param align is requested alignment. If 0, default alignment is used
-  PtrTy salloc(unsigned bytes, uint32_t align = 0);
+  virtual PtrTy salloc(unsigned bytes, uint32_t align = 0) = 0;
 
   /// \brief Allocates memory on the stack and returns a pointer to it
-  PtrTy salloc(Expr elmts, unsigned typeSz, uint32_t align = 0);
+  virtual PtrTy salloc(Expr elmts, unsigned typeSz, uint32_t align = 0) = 0;
 
   /// \brief Returns a pointer value for a given stack allocation
-  PtrTy mkStackPtr(unsigned offset);
+  virtual PtrTy mkStackPtr(unsigned offset) = 0;
 
   /// \brief Pointer to start of the heap
-  PtrTy brk0Ptr();
+  virtual PtrTy brk0Ptr() = 0;
 
   /// \brief Allocates memory on the heap and returns a pointer to it
-  PtrTy halloc(unsigned _bytes, uint32_t align = 0);
+  virtual PtrTy halloc(unsigned _bytes, uint32_t align = 0) = 0;
 
   /// \brief Allocates memory on the heap and returns pointer to it
-  PtrTy halloc(Expr bytes, uint32_t align = 0);
+  virtual PtrTy halloc(Expr bytes, uint32_t align = 0) = 0;
 
   /// \brief Allocates memory in global (data/bss) segment for given global
-  PtrTy galloc(const GlobalVariable &gv, uint32_t align = 0);
+  virtual PtrTy galloc(const GlobalVariable &gv, uint32_t align = 0) = 0;
 
   /// \brief Allocates memory in code segment for the code of a given function
-  PtrTy falloc(const Function &fn);
+  virtual PtrTy falloc(const Function &fn) = 0;
 
   /// \brief Returns a function pointer value for a given function
-  PtrTy getPtrToFunction(const Function &F);
+  virtual PtrTy getPtrToFunction(const Function &F) = 0;
 
   /// \brief Returns a pointer to a global variable
-  PtrTy getPtrToGlobalVariable(const GlobalVariable &gv);
+  virtual PtrTy getPtrToGlobalVariable(const GlobalVariable &gv) = 0;
 
   /// \brief Initialize memory used by the global variable
-  void initGlobalVariable(const GlobalVariable &gv) const;
+  virtual void initGlobalVariable(const GlobalVariable &gv) const = 0;
 
   /// \brief Creates a non-deterministic pointer that is aligned
   ///
   /// Top bits of the pointer are named by \p name and last \c log2(align) bits
   /// are set to zero
-  PtrTy mkAlignedPtr(Expr name, uint32_t align) const;
+  virtual PtrTy mkAlignedPtr(Expr name, uint32_t align) const = 0;
 
   /// \brief Returns sort of a pointer register for an instruction
-  Expr mkPtrRegisterSort(const Instruction &inst) const;
+  virtual Expr mkPtrRegisterSort(const Instruction &inst) const = 0;
 
   /// \brief Returns sort of a pointer register for a function pointer
-  Expr mkPtrRegisterSort(const Function &fn) const;
+  virtual Expr mkPtrRegisterSort(const Function &fn) const = 0;
 
   /// \brief Returns sort of a pointer register for a global pointer
-  Expr mkPtrRegisterSort(const GlobalVariable &gv) const { return ptrSort(); }
+  virtual Expr mkPtrRegisterSort(const GlobalVariable &gv) const = 0;
 
   /// \brief Returns sort of memory-holding register for an instruction
-  Expr mkMemRegisterSort(const Instruction &inst) const;
+  virtual Expr mkMemRegisterSort(const Instruction &inst) const = 0;
 
   /// \brief Returns a fresh aligned pointer value
-  PtrTy freshPtr();
+  virtual PtrTy freshPtr() = 0;
 
   /// \brief Returns a null ptr
-  PtrTy nullPtr() const;
+  virtual PtrTy nullPtr() const = 0;
 
-  /// \brief Pointers have word address (high) and byte offset (low); returns
-  /// number of bits for byte offset
+  /// \brief Pointers have word address (high) and byte offset (low) = 0;
+  /// returns number of bits for byte offset
   ///
   /// \return 0 if unsupported word size
-  unsigned getByteAlignmentBits();
+  virtual unsigned getByteAlignmentBits() = 0;
 
   /// \brief Fixes the type of a havoced value to mach the representation used
   /// by mem repr.
@@ -558,17 +539,7 @@ public:
   /// \param reg
   /// \param val
   /// \return the coerced value.
-  Expr coerce(Expr reg, Expr val);
-
-  /// \brief Symbolic instructions to load a byte from memory, using word
-  /// address and byte address
-  ///
-  /// \param[in] mem memory being accessed
-  /// \param[in] address pointer being accessed, unaligned
-  /// \param[in] offsetBits number of bits at end of pointers reserved for byte
-  ///            address
-  /// \return symbolic value of the byte at the specified address
-  Expr extractUnalignedByte(Expr mem, PtrTy address, unsigned offsetBits);
+  virtual Expr coerce(Expr reg, Expr val) = 0;
 
   /// \brief Loads an integer of a given size from memory register
   ///
@@ -577,53 +548,32 @@ public:
   /// \param[in] byteSz size of the integer in bytes
   /// \param[in] align known alignment of \p ptr
   /// \return symbolic value of the read integer
-  Expr loadIntFromMem(PtrTy ptr, Expr memReg, unsigned byteSz, uint64_t align);
+  virtual Expr loadIntFromMem(PtrTy ptr, Expr memReg, unsigned byteSz,
+                              uint64_t align) = 0;
 
   /// \brief Loads a pointer stored in memory
   /// \sa loadIntFromMem
-  PtrTy loadPtrFromMem(PtrTy ptr, Expr memReg, unsigned byteSz, uint64_t align);
+  virtual PtrTy loadPtrFromMem(PtrTy ptr, Expr memReg, unsigned byteSz,
+                               uint64_t align) = 0;
 
   /// \brief Pointer addition with numeric offset
-  PtrTy ptrAdd(PtrTy ptr, int32_t _offset) const;
+  virtual PtrTy ptrAdd(PtrTy ptr, int32_t _offset) const = 0;
 
   /// \brief Pointer addition with symbolic offset
-  PtrTy ptrAdd(PtrTy ptr, Expr offset) const;
+  virtual PtrTy ptrAdd(PtrTy ptr, Expr offset) const = 0;
 
   /// \brief Stores an integer into memory
   ///
   /// Returns an expression describing the state of memory in \c memReadReg
   /// after the store
   /// \sa loadIntFromMem
-  Expr storeIntToMem(Expr _val, PtrTy ptr, Expr memReadReg, unsigned byteSz,
-                     uint64_t align);
-
-  /// \brief stores integer into memory, address is not word aligned
-  ///
-  /// \sa storeIntToMem
-  Expr storeUnalignedIntToMem(Expr val, PtrTy ptr, Expr mem, unsigned byteSz);
-
-  /// \brief Given a word, updates a byte
-  ///
-  /// \param word existing word
-  /// \param byteData updated byte
-  /// \param byteOffset symbolic pointer indicating which byte to update
-  /// \return updated word
-  Expr setByteOfWord(Expr word, Expr byteData, PtrTy byteOffset);
+  virtual Expr storeIntToMem(Expr _val, PtrTy ptr, Expr memReadReg,
+                             unsigned byteSz, uint64_t align) = 0;
 
   /// \brief Stores a pointer into memory
   /// \sa storeIntToMem
-  Expr storePtrToMem(PtrTy val, PtrTy ptr, Expr memReadReg, unsigned byteSz,
-                     uint64_t align);
-
-  /// \brief Creates bit-vector of a given width filled with 0
-  Expr mkZeroE(unsigned width, ExprFactory &efac) {
-    return bv::bvnum(0UL, width, efac);
-  }
-
-  /// brief Creates a bit-vector for number 1 of a given width
-  Expr mkOneE(unsigned width, ExprFactory &efac) {
-    return bv::bvnum(1UL, width, efac);
-  }
+  virtual Expr storePtrToMem(PtrTy val, PtrTy ptr, Expr memReadReg,
+                             unsigned byteSz, uint64_t align) = 0;
 
   /// \brief Returns an expression corresponding to a load from memory
   ///
@@ -631,71 +581,77 @@ public:
   /// \param[in] memReg is the memory register being read
   /// \param[in] ty is the type of value being loaded
   /// \param[in] align is the known alignment of the load
-  Expr loadValueFromMem(PtrTy ptr, Expr memReg, const llvm::Type &ty,
-                        uint64_t align);
+  virtual Expr loadValueFromMem(PtrTy ptr, Expr memReg, const llvm::Type &ty,
+                                uint64_t align) = 0;
 
-  Expr storeValueToMem(Expr _val, PtrTy ptr, Expr memReadReg, Expr memWriteReg,
-                       const llvm::Type &ty, uint32_t align);
+  virtual Expr storeValueToMem(Expr _val, PtrTy ptr, Expr memReadReg,
+                               Expr memWriteReg, const llvm::Type &ty,
+                               uint32_t align) = 0;
 
   /// \brief Executes symbolic memset with a concrete length
-  Expr MemSet(PtrTy ptr, Expr _val, unsigned len, Expr memReadReg,
-              Expr memWriteReg, uint32_t align);
+  virtual Expr MemSet(PtrTy ptr, Expr _val, unsigned len, Expr memReadReg,
+                      Expr memWriteReg, uint32_t align) = 0;
 
   /// \brief Executes symbolic memcpy with concrete length
-  Expr MemCpy(PtrTy dPtr, PtrTy sPtr, unsigned len, Expr memTrsfrReadReg,
-              Expr memReadReg, Expr memWriteReg, uint32_t align);
+  virtual Expr MemCpy(PtrTy dPtr, PtrTy sPtr, unsigned len,
+                      Expr memTrsfrReadReg, Expr memReadReg, Expr memWriteReg,
+                      uint32_t align) = 0;
 
   /// \brief Executes symbolic memcpy from physical memory with concrete length
-  Expr MemFill(PtrTy dPtr, char *sPtr, unsigned len, uint32_t align = 0);
+  virtual Expr MemFill(PtrTy dPtr, char *sPtr, unsigned len,
+                       uint32_t align = 0) = 0;
 
   /// \brief Executes inttoptr conversion
-  PtrTy inttoptr(Expr intVal, const Type &intTy, const Type &ptrTy) const;
+  virtual PtrTy inttoptr(Expr intVal, const Type &intTy, const Type &ptrTy) const = 0;
 
   /// \brief Executes ptrtoint conversion
-  Expr ptrtoint(PtrTy ptr, const Type &ptrTy, const Type &intTy) const;
+  virtual Expr ptrtoint(PtrTy ptr, const Type &ptrTy,
+                        const Type &intTy) const = 0;
 
-  Expr ptrUlt(PtrTy p1, PtrTy p2) const;
-  Expr ptrSlt(PtrTy p1, PtrTy p2) const;
-  Expr ptrUle(PtrTy p1, PtrTy p2) const;
-  Expr ptrSle(PtrTy p1, PtrTy p2) const;
-  Expr ptrUgt(PtrTy p1, PtrTy p2) const;
-  Expr ptrSgt(PtrTy p1, PtrTy p2) const;
-  Expr ptrUge(PtrTy p1, PtrTy p2) const;
-  Expr ptrSge(PtrTy p1, PtrTy p2) const;
+  virtual Expr ptrUlt(PtrTy p1, PtrTy p2) const = 0;
+  virtual Expr ptrSlt(PtrTy p1, PtrTy p2) const = 0;
+  virtual Expr ptrUle(PtrTy p1, PtrTy p2) const = 0;
+  virtual Expr ptrSle(PtrTy p1, PtrTy p2) const = 0;
+  virtual Expr ptrUgt(PtrTy p1, PtrTy p2) const = 0;
+  virtual Expr ptrSgt(PtrTy p1, PtrTy p2) const = 0;
+  virtual Expr ptrUge(PtrTy p1, PtrTy p2) const = 0;
+  virtual Expr ptrSge(PtrTy p1, PtrTy p2) const = 0;
 
   /// \brief Checks if two pointers are equal.
-  Expr ptrEq(PtrTy p1, PtrTy p2) const;
-  Expr ptrNe(PtrTy p1, PtrTy p2) const;
+  virtual Expr ptrEq(PtrTy p1, PtrTy p2) const = 0;
+  virtual Expr ptrNe(PtrTy p1, PtrTy p2) const = 0;
+
+  virtual Expr ptrSub(PtrTy p1, PtrTy p2) const = 0;
+
+  /// \brief Computes a pointer corresponding to the gep instruction
+  virtual PtrTy gep(PtrTy ptr, gep_type_iterator it,
+                    gep_type_iterator end) const = 0;
+
+  /// \brief Called when a function is entered for the first time
+  virtual void onFunctionEntry(const Function &fn) = 0;
+
+  /// \brief Called when a module entered for the first time
+  virtual void onModuleEntry(const Module &M) = 0;
+
+  /// \brief Debug helper
+  virtual void dumpGlobalsMap() = 0;
+
+  virtual std::pair<char *, unsigned>
+  getGlobalVariableInitValue(const llvm::GlobalVariable &gv) = 0;
 
   /// \brief Checks if \p a <= b <= c.
   Expr ptrInRangeCheck(PtrTy a, PtrTy b, PtrTy c) {
     return mk<AND>(ptrUle(a, b), ptrUle(b, c));
   }
-
-  Expr ptrSub(PtrTy p1, PtrTy p2) const;
-
   /// \brief Calculates an offset of a pointer from its base.
   Expr ptrOffsetFromBase(PtrTy base, PtrTy ptr) { return ptrSub(ptr, base); }
 
-  /// \brief Computes a pointer corresponding to the gep instruction
-  PtrTy gep(PtrTy ptr, gep_type_iterator it, gep_type_iterator end) const;
-
-  /// \brief Called when a function is entered for the first time
-  void onFunctionEntry(const Function &fn);
-
-  /// \brief Called when a module entered for the first time
-  void onModuleEntry(const Module &M);
-
-  /// \brief Debug helper
-  void dumpGlobalsMap() { return m_allocator->dumpGlobalsMap(); }
-
-  std::pair<char *, unsigned>
-  getGlobalVariableInitValue(const llvm::GlobalVariable &gv) {
-    return m_allocator->getGlobalVariableInitValue(gv);
-  }
-
   uint32_t getAlignment(const llvm::Value &v) const { return m_alignment; }
 };
+
+OpSemMemManager *mkRawMemManager(Bv2OpSem &sem, Bv2OpSemContext &ctx,
+                                 unsigned ptrSz, unsigned wordSz,
+                                 bool useLambdas = false);
 
 /// \Brief Base class for memory representation
 class OpSemMemRepr {
@@ -797,8 +753,8 @@ class ConstantExprEvaluator {
     assert(m_ctx);
     return *m_ctx;
   }
-  /// \brief Stores a value in \p Val to memory pointed by \p Ptr. The store is
-  /// of type \p Ty
+  /// \brief Stores a value in \p Val to memory pointed by \p Ptr. The store
+  /// is of type \p Ty
   void storeValueToMemory(const GenericValue &Val, GenericValue *Ptr, Type *Ty);
 
 public:
